refactored backend services to combine kg and db into a single interface
