+++ backend/src/kg/db/models.py
@@ -1,6 +1,7 @@
 from neomodel import (
     AsyncStructuredNode, StringProperty, IntegerProperty, UniqueIdProperty,
-    JSONProperty, AsyncRelationshipTo, AsyncRelationshipFrom, ZeroOrOne, ZeroOrMore
+    JSONProperty, AsyncRelationshipTo, AsyncRelationshipFrom, ZeroOrOne, ZeroOrMore,
+    ArrayProperty, FloatProperty, VectorIndex
 )
 
 class Author(AsyncStructuredNode):
@@ -60,6 +61,10 @@ class Paper(AsyncStructuredNode):
     influential_citation_count = IntegerProperty()
     publication_types = JSONProperty(default=[])
     publication_date = StringProperty()
+    abstract_embedding = ArrayProperty(
+        FloatProperty(),
+        vector_index=VectorIndex(dimensions=768, similarity_function='cosine')
+    )
     
     # Relationships
     references = AsyncRelationshipFrom("Paper", "CITES")
@@ -93,8 +98,29 @@ class ChatResponse(AsyncStructuredNode):
     message = StringProperty()
     response_to = AsyncRelationshipTo("ChatMessage", "RESPONSE_TO")
 
-class Workspace(AsyncStructuredNode):
+class Chunk(AsyncStructuredNode):
+    uid = UniqueIdProperty()
+    chunkId = StringProperty(unique_index=True, required=True)
+    paper_id = StringProperty()
+    source = StringProperty()
+    text = StringProperty()
+    textEmbedding = ArrayProperty(
+        FloatProperty(),
+        vector_index=VectorIndex(dimensions=768, similarity_function='cosine')
+    )
+    
+    # Relationships
+    paper = AsyncRelationshipTo("Paper", "BELONGS_TO_PAPER")
+    project = AsyncRelationshipTo("Project", "BELONGS_TO")
+    next_chunk = AsyncRelationshipTo("Chunk", "NEXT")
+    previous_chunk = AsyncRelationshipFrom("Chunk", "NEXT")
+
+class Project(AsyncStructuredNode):
     uid = UniqueIdProperty()
     name = StringProperty(unique_index=True, required=True)
-    papers = AsyncRelationshipFrom("Paper", "BELONGS_TO")
-    authors = AsyncRelationshipFrom("Author", "BELONGS_TO")
+    papers = AsyncRelationshipFrom("Paper", "BELONGS_TO", cardinality=ZeroOrMore)
+    authors = AsyncRelationshipFrom("Author", "BELONGS_TO", cardinality=ZeroOrMore)
+    documents = AsyncRelationshipFrom("Document", "BELONGS_TO", cardinality=ZeroOrMore)
+    tags = AsyncRelationshipFrom("Tag", "BELONGS_TO", cardinality=ZeroOrMore)
+    journals = AsyncRelationshipFrom("Journal", "BELONGS_TO", cardinality=ZeroOrMore)
+    publicationVenues = AsyncRelationshipFrom("PublicationVenue", "BELONGS_TO", cardinality=ZeroOrMore)
