+++ backend/src/kg/db/util.py
