+++ docker-compose.yml
@@ -70,26 +70,6 @@ services:
     depends_on:
       - nexarag.api
 
-  nexarag.db:
-    image: nexarag.db:latest
-    build:
-      context: ./backend/src
-      dockerfile: db/Dockerfile
-    container_name: nexarag.db
-    restart: unless-stopped
-    environment:
-      NEO4J_URI: bolt://nexarag.neo4j:7687
-      NEO4J_USERNAME: neo4j
-      NEO4J_PASSWORD: password
-      NEO4J_DATABASE: neo4j
-      RABBITMQ_PASSWORD: guest
-      RABBITMQ_USERNAME: guest
-    depends_on:
-      nexarag.neo4j:
-        condition: service_healthy
-      nexarag.rabbitmq:
-        condition: service_healthy
-
   nexarag.kg:
     image: nexarag.kg:latest
     build:
